backend/requirements.txt
flask
google-cloud-vision
google-generativeai
python-dotenv
requests
flask-cors
gunicorn